+++ purchase_fiscal_position_update/__manifest__.py
@@ -1,28 +1,12 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Purchase Fiscal Position Update module for Odoo
-#    Copyright (C) 2011-2014 Julius Network Solutions SARL <contact@julius.fr>
-#    Copyright (C) 2014 Akretion (http://www.akretion.com)
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# © 2011-2014 Julius Network Solutions SARL <contact@julius.fr>
+# © 2014-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
 
 {
     'name': 'Purchase Fiscal Position Update',
-    'version': '8.0.1.0.0',
+    'version': '10.0.1.0.0',
     'category': 'Purchase Management',
     'license': 'AGPL-3',
     'summary': 'Changing the fiscal position of a purchase order will '
@@ -40,6 +24,5 @@ updated and a warning is displayed to the user in this case.
               "Akretion,"
               "Odoo Community Association (OCA)",
     'depends': ['purchase'],
-    'data': [],
-    'installable': False,
+    'installable': True,
 }
